# packages/shared-contracts/src/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "__Host-quart-admin-session": {
        "description": "Better Auth session cookie for the admin console (HTTP-only).",
        "in": "cookie",
        "name": "__Host-quart-admin-session",
        "type": "apiKey"
      },
      "__Host-quart-api-session": {
        "description": "Better Auth session cookie for the mobile PWA (HTTP-only).",
        "in": "cookie",
        "name": "__Host-quart-api-session",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "EdDSA",
        "description": "Ed25519-signed JWT issued by /auth/login.",
        "scheme": "bearer",
        "type": "http"
      }
    },
...
  },
  "info": {
    "description": "Mobile + Admin shared API",
    "title": "Quart API",
    "version": "0.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/admin/ideas/{id}/comments": {
      "post": {
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin/ideas"
        ]
      }
    },
...
    "/admin/roles": {
      "get": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin/roles"
        ]
      },
      "post": {
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin/roles"
        ]
      }
    },
    "/admin/roles/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin/roles"
        ]
      },
      "patch": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin/roles"
        ]
      }
    },
...
    "/ideas": {
      "get": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      },
      "post": {
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      }
    },
    "/ideas/{id}": {
      "delete": {
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      },
      "put": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      }
    },
    "/ideas/{id}/comments": {
      "get": {
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      },
      "post": {
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      }
    },
    "/ideas/{id}/vote": {
      "delete": {
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      },
      "post": {
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ideas"
        ]
      }
    },
...
  },
  "servers": [],
...
}